Added authentication support

+++ services/notes-storage.mjs
@@ -3,7 +3,7 @@
 import Datastorage from 'nedb-promise';
 
 class Note {
-    constructor (creationDate, title, description, importance, dueDate, completionDate, completed) {
+    constructor (creationDate, title, description, importance, dueDate, completionDate, completed, user) {
         this.creationDate = creationDate || new Date();
         this.title = title;
         this.description = description;
@@ -13,6 +13,7 @@ class Note {
         this.dueDate = dueDate || defaultDueDate.toISOString().slice(0,10);
         this.completionDate = completionDate;
         this.completed = completed;
+        this.user = user;
     }
 }
 
@@ -22,26 +23,26 @@ class NotesStorage {
         this.db = db || new Datastorage({filename: './data/notes.db', autoload: true});
     }
 
-    async getAll() {
-        return await this.db.find({});
+    async getAll(currentUser) {
+        return await this.db.find({user: currentUser});
     }
 
-    async getById(id) {
-        return await this.db.findOne({_id: id});
+    async getById(id, currentUser) {
+        return await this.db.findOne({_id: id, user: currentUser});
     }
 
-    async add(creationDate, title, description, importance, dueDate, completionDate, completed) {
-        let note = new Note(creationDate, title, description, importance, dueDate, completionDate, completed);
+    async add(creationDate, title, description, importance, dueDate, completionDate, completed, currentUser) {
+        const note = new Note(creationDate, title, description, importance, dueDate, completionDate, completed, currentUser);
         return await this.db.insert(note);
     }
 
-    async update(id, changes) {
-        await this.db.update({_id: id}, {$set: changes});
+    async update(id, changes, currentUser) {
+        await this.db.update({_id: id, user: currentUser}, {$set: changes});
         return await this.getById(id);
     }
 
-    async delete(id) {
-        await this.db.remove({_id: id});
+    async delete(id, currentUser) {
+        await this.db.remove({_id: id, user: currentUser});
         return true;
     }
